# docs/evidence/TASK-SD-006-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "IndustrialPlatform.SystemData.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/service-initialization/registrations/{serviceKey}/{moduleKey}": {
      "put": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "serviceKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "moduleKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationManifestV2"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationManifestV2"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationManifestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "serviceKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "moduleKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistrationV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/registrations": {
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "serviceKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "moduleKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseRegistrationSummaryV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseRegistrationSummaryV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseRegistrationSummaryV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/plans": {
      "post": {
        "tags": [
          "ServiceInitialization"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationPlanRequestV2"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationPlanRequestV2"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationPlanRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabasePlanV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabasePlanV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabasePlanV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/plans/{planNId}": {
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "planNId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlanV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlanV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlanV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/plans/{planNId}/approvals": {
      "post": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "planNId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApprovalRequestV1"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApprovalRequestV1"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApprovalRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApprovalV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApprovalV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApprovalV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/plans/{planNId}/backup-evidence": {
      "post": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "planNId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseBackupEvidenceRequestV1"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseBackupEvidenceRequestV1"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseBackupEvidenceRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBackupEvidenceV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBackupEvidenceV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBackupEvidenceV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/operations/apply": {
      "post": {
        "tags": [
          "ServiceInitialization"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationApplyRequestV2"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationApplyRequestV2"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitializationApplyRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueOperationV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/operations": {
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseOperationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseOperationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfDatabaseOperationV1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-initialization/operations/{operationNId}": {
      "get": {
        "tags": [
          "ServiceInitialization"
        ],
        "parameters": [
          {
            "name": "operationNId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseOperationV1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseOperationV1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseOperationV1"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceInitializationManifestV2": {
        "type": "object",
        "properties": {
          "serviceKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "moduleKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "logicalDatabaseName": {
            "type": [
              "null",
              "string"
            ]
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "topologyMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "migrationArtifactId": {
            "type": [
              "null",
              "string"
            ]
          },
          "requestedVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "artifactChecksum": {
            "type": [
              "null",
              "string"
            ]
          },
          "artifactSignature": {
            "type": [
              "null",
              "string"
            ]
          },
          "desiredState": {
            "type": [
              "null",
              "string"
            ]
          },
          "autoProvision": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "autoMigrate": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "ownerNId": {
            "type": [
              "null",
              "string"
            ]
          },
          "manifestVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "seedSets": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/SeedSetV1"
            }
          }
        }
      },
...
    }
  },
...
}